+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # next
 
+* Call DartPluginRegistrant.ensureInitialized when isolate is starting
 * Documentation and example update to fix (#374) WorkManager not working when App is obfuscated or using Flutter 3.1+
 
 # 0.5.0